+++ src/components/resources-side-panel/Resources.jsx
@@ -1,25 +1,17 @@
-import {getLanguageClass, LANGUAGE, mapLanguageCode, menuItems} from "../../utils/Constants.js";
+import {LANGUAGE, mapLanguageCode, menuItems} from "../../utils/Constants.js";
 import axiosInstance from "../../config/axios-config.js";
 import {useQuery} from "react-query";
-import {IoMdCheckmark, IoMdClose} from "react-icons/io";
-import {IoCopy, IoLanguage, IoNewspaperOutline} from "react-icons/io5";
-import {BsFacebook, BsTwitter, BsWindowFullscreen} from "react-icons/bs";
+import {IoMdClose} from "react-icons/io";
+import {IoLanguage, IoNewspaperOutline} from "react-icons/io5";
+import {BsWindowFullscreen} from "react-icons/bs";
 import {FiInfo, FiList} from "react-icons/fi";
 import {BiBook, BiSearch} from "react-icons/bi";
 import {useState} from "react";
 import {useTranslate} from "@tolgee/react";
-import "./Resources.scss"
-import { GoLinkExternal } from "react-icons/go";
-export const fetchtranslationdata=async(segment_id,skip=0,limit=10)=>{
-  const {data} = await axiosInstance.get(`/api/v1/segments/${segment_id}/translations`, {
-    params: {
-      segment_id,
-      skip,
-      limit
-    }
-  });
-  return data;
-}
+import "./Resources.scss";
+import ShareView from "./components/ShareView";
+import TranslationView from "./components/TranslationView";
+
 export const fetchSidePanelData = async (text_id) => {
   const storedLanguage = localStorage.getItem(LANGUAGE);
   const language = (storedLanguage ? mapLanguageCode(storedLanguage) : "bo");
@@ -34,7 +26,6 @@ export const fetchSidePanelData = async (text_id) => {
 const Resources = ({textId, segmentId, showPanel, setShowPanel, setVersionId, versionId}) => {
   const [isShareView, setIsShareView] = useState(false);
   const [isTranslationView, setIsTranslationView] = useState(false);
-  const [copied, setCopied] = useState(false);
   const [expandedTranslations, setExpandedTranslations] = useState({});
 
   const {t} = useTranslate();
@@ -47,140 +38,26 @@ const Resources = ({textId, segmentId, showPanel, setShowPanel, setVersionId, ve
       staleTime: 1000 * 60 * 20
     }
   );
-  const {data: sidepaneltranslation} = useQuery(
-    ["sidePaneltranslation",segmentId],
-    () => fetchtranslationdata("2353849b-f8fa-43e4-850d-786b623d0130"), //send segmentId later todo
-    {
-      refetchOnWindowFocus: false,
-      staleTime: 1000 * 60 * 20
-    }
-  );
 
-  const languageMap = {
-    "zh": "language.sanskrit",
-    "bo": "language.tibetan",
-    "en": "language.english",
-    "ja": "language.japanese",
-    "ko": "language.korean",
-    "fr": "language.french",
-    "de": "language.german",
-    "bhu":"language.bhutanese",
-    "mo":"language.mongolian",
-    "sp":"language.spanish"
-  }
   const renderShareView = () => {
     return (
-      <div>
-        <div className="headerthing">
-          <p className='mt-4 px-4 listtitle'>{t('panel.resources')}</p>
-          <IoMdClose
-            size={24}
-            onClick={() => setIsShareView(false)}
-            className="close-icon"
-          />
-        </div>
-        <div className="share-content p-3">
-          <p className="mb-3 textgreat ">{t('text.share_link')}</p>
-          <div className="share-url-container p-3 mb-3">
-            <p className="share-url text-truncate">{sidePanelData?.text_infos?.short_url}</p>
-            <button
-              className="copy-button"
-              onClick={() => {
-                navigator.clipboard.writeText(sidePanelData?.text_infos?.short_url);
-                setCopied(true);
-                setTimeout(() => {
-                  setCopied(false);
-                }, 3000);
-              }}
-            >
-              {copied ? <IoMdCheckmark size={16}/> : <IoCopy size={16}/>}
-            </button>
-          </div>
-          <p className="textgreat">{t('text.more_options')}</p>
-          <div className="social-share-buttons">
-            <p className="social-button">
-              <BsFacebook className="social-icon"/>{t('common.share_on_fb')}
-            </p>
-            <p className="social-button">
-              <BsTwitter className="social-icon"/>{t('common.share_on_x')}
-            </p>
-          </div>
-        </div>
-      </div>
+      <ShareView 
+        sidePanelData={sidePanelData} 
+        setIsShareView={setIsShareView} 
+      />
     );
   };
 
   const renderTranslationView = () => {
-    const groupedTranslations = sidepaneltranslation?.translations?.reduce((acc, translation) => {
-      if (!acc[translation.language]) {
-        acc[translation.language] = [];
-      }
-      acc[translation.language].push(translation);
-      return acc;
-    }, {});
     return (
-      <div>
-        <div className="headerthing">
-          <p className='mt-4 px-4 listtitle'>{t('connection_pannel.translations')}</p>
-          <IoMdClose
-            size={24}
-            onClick={() => setIsTranslationView(false)}
-            className="close-icon"
-          />
-        </div>
-        <div className="translation-content p-4">
-          <div className="translations-list">
-            {groupedTranslations && Object.entries(groupedTranslations).map(([language, translations]) => (
-              <div key={language} className="language-group">
-                <h3 className="language-title navbaritems">
-                  {t(languageMap[language])} <span className="translation-count">({translations.length})</span>
-                </h3>
-                {translations.map((translation, index) => {
-                  const translationKey = `${language}-${index}`;
-                  const isExpanded = expandedTranslations[translationKey] || false;
-                  
-                  return (
-                    <div key={index} className="translation-item">
-                      <span className={`translation-content ${getLanguageClass(translation.language)}`}>
-                        <div 
-                          className={`translation-text ${isExpanded ? 'expanded' : 'collapsed'}`}
-                          dangerouslySetInnerHTML={{__html: translation.content}} 
-                        />
-                        {translation.content && translation.content.length > 0 && (
-                          <button 
-                            className="expand-button "
-                            onClick={() => setExpandedTranslations(prev => ({
-                              ...prev,
-                              [translationKey]: !isExpanded
-                            }))}
-                          >
-                            {isExpanded ? t('panel.showless') : t('panel.showmore')}
-                          </button>
-                        )}
-                      </span>
-                      <div className={` belowdiv ${getLanguageClass(translation.language)}`}>                    
-                      <p className=" titles"> {translation?.title ? translation.title : ""}</p>                 
-                     {translation.source && <p className="navbaritems"> {t("connection_panel.menuscript.source")}: { translation.source}</p> }   
-                         
-                         <p className="textgreat review navbaritems">{t("text.versions.information.review_history")}</p> 
-                          <div className=" linkselect navbaritems">
-                            <div className="linkicons ">
-                            <GoLinkExternal/>
-                            {t("text.translation.open_text")}
-                            </div>
-                            <p onClick={()=>setVersionId(translation.text_id)} className="selectss navbaritems">
-                              {translation.text_id === versionId ? t("text.translation.current_selected") : t("common.select")}
-                            </p>
-                          </div>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
+      <TranslationView 
+        segmentId={segmentId}
+        setIsTranslationView={setIsTranslationView}
+        expandedTranslations={expandedTranslations}
+        setExpandedTranslations={setExpandedTranslations}
+        setVersionId={setVersionId}
+        versionId={versionId}
+      />
     );
   };
