+++ src/components/resources-side-panel/Resources.scss
@@ -7,91 +7,10 @@
   box-shadow: -2px 0 5px rgba(0, 0, 0, 0.05);
   height: 100%;
 
-  .share-content {
-    text-align: start;
-    .textgreat{
-      color: #828282;
-      border-bottom: 1px solid #f0f0f0;
-    }
-
-    .social-share-buttons {
-      margin-top: 16px;
-
-      .social-button {
-        display: flex;
-        align-items: center;
-        padding: 12px 16px;
-        margin: 8px 0;
-        cursor: pointer;
-        transition: background-color 0.2s ease;
-        border-radius: 4px;
-
-        &:hover {
-          background-color: #f5f5f5;
-        }
-
-        .social-icon {
-          color: #666;
-          margin-right: 12px;
-          font-size: 18px;
-        }
-      }
-    }
-
-    .share-url-container {
-      display: flex;
-      align-items: center;
-      background-color: #f5f5f5;
-      border-radius: 4px;
-      gap: 10px;
-
-      .share-url {
-        flex: 1;
-        margin: 0;
-        font-size: 14px;
-        color: #666;
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-      }
-
-      .copy-button {
-        background: none;
-        border: none;
-        padding: 5px;
-        cursor: pointer;
-        color: #666;
-
-        &:hover {
-          color: #333;
-        }
-      }
-    }
-  }
-
   &.show {
     width: 400px;
   }
-  .headerthing{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background-color: #f0f0f0;
-  }
-  .close-icon {
-    cursor: pointer;
-    padding: 4px;
-    border-radius: 50%;
-    background-color: #f0f0f0;
-    color: #333;
-    margin-right: 10px;
-    transition: all 0.2s ease-in-out;
 
-    &:hover {
-      background-color: #e0e0e0;
-      transform: scale(1.1);
-    }
-  }
   .panel-content {
     text-align: start;
     opacity: 0;
@@ -120,133 +39,11 @@
         color: #7b7b7b;
         background-color: #f8f9fa;
       }
-
-
     }
-
-
   }
-
 
   &.show .panel-content {
     opacity: 1;
-  }
-
-  .translation-content {
-    text-align: start;
-    font-size: medium;
-    line-height: 2;
-    color: black;
-    
-    .translation-text {
-      position: relative;
-      
-      &.collapsed {
-        max-height: 6em;
-        overflow: hidden;
-      }
-      
-      &.expanded {
-        max-height: none;
-      }
-    }
-    
-    .expand-button {
-      background: none;
-      border: none;
-      color: #C74444;
-      cursor: pointer;
-      
-      margin-top: 4px;
-      font-size: 14px;
-      font-weight: 500;
-      
-      &:hover {
-        text-decoration: underline;
-      }
-    }
-    
-    .belowdiv{
-      margin-top: 10px;
-      padding: 1rem;
-      border-radius: 6px;
-      background-color: #FBFBFA;
-      font-size: small;
-      .titles{
-        font-size: medium;
-        font-weight: 600;
-        line-height: 1;
-      }
-      p ,div{
-        font-size: small;
-        margin: 0;
-        padding: 2px 0;
-        line-height: 1;
-      }
-      .review{
-        margin-top: 10px;
-      }
-      .linkselect{
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        
-        min-height: 40px;
-
-        .linkicons{
-          display: flex;
-          gap: 0px 6px;
-          align-items: center;
-          height: 100%;
-        }
-        .selectss{
-          color: #0B71E7;
-          margin: 0;
-          padding: 0;
-          display: flex;
-          align-items: center;
-          height: 100%;
-          cursor: pointer;
-        }
-      }
-    }
-    .textgreat {
-      color: #828282;
-      border-bottom: 1px solid #f0f0f0;
-      margin-bottom: 12px;
-    }
-
-    .segment-content {
-      background-color: #f5f5f5;
-      border-radius: 4px;
-      font-size: 14px;
-      color: #333;
-      margin-bottom: 20px;
-    }
-
-    .translations-list {
-      .language-group {
-        margin-bottom: 1rem;
-        .language-title {
-          padding-bottom: 1rem;
-          border-bottom: #C74444 solid 2px;
-          margin-bottom: 0.75rem;
-          color: #7d7d7d;
-
-          .translation-count {
-           
-            color: #718096;
-            font-size: 0.9rem;
-            margin-left: 0.25rem;
-          }
-        }
-      }
-
-      .translation-item {
-        margin-bottom: 0.5rem;
-        border-radius: 0.375rem;
-      }
-    }
   }
 }
