+++ src/components/resources-side-panel/components/ShareView.scss
@@ -1,0 +1,83 @@
+.share-content {
+  text-align: start;
+  .textgreat{
+    color: #828282;
+    border-bottom: 1px solid #f0f0f0;
+  }
+
+  .social-share-buttons {
+    margin-top: 16px;
+
+    .social-button {
+      display: flex;
+      align-items: center;
+      padding: 12px 16px;
+      margin: 8px 0;
+      cursor: pointer;
+      transition: background-color 0.2s ease;
+      border-radius: 4px;
+
+      &:hover {
+        background-color: #f5f5f5;
+      }
+
+      .social-icon {
+        color: #666;
+        margin-right: 12px;
+        font-size: 18px;
+      }
+    }
+  }
+
+  .share-url-container {
+    display: flex;
+    align-items: center;
+    background-color: #f5f5f5;
+    border-radius: 4px;
+    gap: 10px;
+
+    .share-url {
+      flex: 1;
+      margin: 0;
+      font-size: 14px;
+      color: #666;
+      overflow: hidden;
+      white-space: nowrap;
+      text-overflow: ellipsis;
+    }
+
+    .copy-button {
+      background: none;
+      border: none;
+      padding: 5px;
+      cursor: pointer;
+      color: #666;
+
+      &:hover {
+        color: #333;
+      }
+    }
+  }
+}
+
+.headerthing {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: #f0f0f0;
+}
+
+.close-icon {
+  cursor: pointer;
+  padding: 4px;
+  border-radius: 50%;
+  background-color: #f0f0f0;
+  color: #333;
+  margin-right: 10px;
+  transition: all 0.2s ease-in-out;
+
+  &:hover {
+    background-color: #e0e0e0;
+    transform: scale(1.1);
+  }
+}
